+++ src/main/resources/docs/CHANGELOG.md
@@ -7,6 +7,8 @@ This log is for tracking changes that may not be pertinent to the individual too
 Changes
 -------
 
+- **2026-05-10** DoomMake preference drop-downs did not set the correct value for success/failure sounds when opened (but does on change).
+
 - **2026-05-09** Fixes for File fields and copy-pasting.
 
 - **2026-03-26** Better detection for if DoomTools GUI is already running (uses temp file channels instead of socket ports).